+++ src/app/app-routing.module.ts
@@ -1,11 +1,12 @@
 import {NgModule} from '@angular/core';
 import {
-  LoadChildrenCallback,
   ROUTES,
   RouterModule,
   Routes,
 } from '@angular/router';
 import {APP_CONFIG, AppConfig} from './core/config/app-config';
+import {PageNotFoundComponent} from "./component-library/error-404-component/page-not-found.component";
+
 
 @NgModule({
   imports: [RouterModule.forRoot([], {paramsInheritanceStrategy: 'always'})],
@@ -15,27 +16,31 @@ import {APP_CONFIG, AppConfig} from './core/config/app-config';
       provide: ROUTES,
       deps: [APP_CONFIG],
       multi: true,
-      useFactory: (config: AppConfig): Routes => {
-        const loadChildRoutes = (
-          loadChildren: LoadChildrenCallback,
-        ): Routes => [{path: '', loadChildren}];
 
+      useFactory: (config: AppConfig): Routes => {
+        const routes: Routes = [];
         switch (config.routes) {
           case 'broker-ui':
-            return loadChildRoutes(() =>
-              import('./routes/broker-ui/broker-ui.module').then(
-                (m) => m.BrokerUiModule,
-              ),
-            );
+            routes.push({
+              path: '', loadChildren: () =>
+                import('./routes/broker-ui/broker-ui.module').then(
+                  (m) => m.BrokerUiModule,
+                )
+            });
+            break;
           case 'connector-ui':
-            return loadChildRoutes(() =>
-              import('./routes/connector-ui/connector-ui.module').then(
-                (m) => m.ConnectorUiModule,
-              ),
-            );
+            routes.push({
+              path: '', loadChildren: () =>
+                import('./routes/connector-ui/connector-ui.module').then(
+                  (m) => m.ConnectorUiModule,
+                )
+            });
+            break;
           default:
             throw new Error(`Unhandled PageSet: ${config.routes}`);
         }
+        routes.push({path: '**', component: PageNotFoundComponent})
+        return routes;
       },
     },
   ],