+++ src/app/component-library/catalog/asset-detail-dialog/asset-property-grid-group-builder.ts
@@ -50,7 +50,6 @@ export class AssetPropertyGridGroupBuilder {
         labelTitle: AssetProperties.language,
         ...this.propertyGridUtils.guessValue(asset.language?.label),
       },
-
       {
         icon: 'apartment',
         label: 'Publisher',